talk_to_image/pyproject.toml
[tool.poetry]
name = "image_processing"
version = "0.1.0"
description = ""
authors = ["Pramod Modi <pramodkmodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.3.13"
langchain-community = "^0.3.13"
torch = "2.2.2"
langchain-huggingface = "^0.1.2"
sentence-transformers = "^3.3.1"
pytesseract = "^0.3.13"
unstructured-client = "^0.28.1"
unstructured = {extras = ["all-docs"], version = "^0.16.11"}
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.9.0.post1"
langchain-mistralai = "^0.2.4"
streamlit = "^1.41.1"
streamlit-chat = "^0.1.1"
pillow = "^11.1.0"
tavily-python = "^0.5.0"
langgraph = "^0.2.61"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
